requirements.txt
# INTELLIGIT BOT Requirements - ADVANCED ENSEMBLE SYSTEM
# Core dependencies for the trading bot with ML/DL capabilities

# IQ Option API with compatible websocket version
iqoptionapi>=0.5
websocket-client==0.56.0

# Data processing and analysis
pandas>=1.3.0
numpy>=1.20.0
scipy>=1.7.0

# Technical analysis
ta>=0.7.0

# Machine Learning - Core
scikit-learn>=1.0.0
xgboost>=1.5.0
lightgbm>=3.3.0

# Deep Learning - PyTorch (Optional but recommended)
torch>=1.10.0
torchvision>=0.11.0

# Deep Learning - TensorFlow (Alternative)
tensorflow>=2.8.0

# Feature Engineering and Statistics
statsmodels>=0.13.0

# HTTP requests for Telegram
requests>=2.25.0

# Async support
aiohttp>=3.8.0

# Logging and utilities
python-dotenv>=0.19.0

# API server (dashboard)
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
requests>=2.31.0

# Data visualization (optional)
matplotlib>=3.5.0
seaborn>=0.11.0

# Performance optimization
numba>=0.56.0

# Memory management
psutil>=5.8.0

# Machine learning dependencies
scikit-learn>=1.0.0
xgboost>=1.5.0
torch>=1.12.0
torchvision>=0.13.0
tensorflow>=2.8.0
keras>=2.8.0
# torch>=1.12.0
# scikit-learn>=1.0.0

# Optional: For advanced features
# matplotlib>=3.3.0
# seaborn>=0.11.0
